refactor: Naming the walker
