+++ .prettierignore
@@ -1,1 +1,3 @@
 package.json
+test/code/
+lib/
