+++ src/internal/OpenApiTools/Walker/Definition.ts
