+++ src/internal/OpenApiTools/Walker/Operation.ts
