+++ src/internal/OpenApiTools/Walker/State.ts
