+++ src/internal/OpenApiTools/Walker/Store.ts
