+++ src/internal/OpenApiTools/Walker/index.ts
