+++ src/internal/OpenApiTools/Walker/structure/DataStructure.ts
