+++ src/internal/OpenApiTools/Walker/structure/InterfaceNode.ts
