+++ src/internal/OpenApiTools/Walker/structure/NamespaceTree.ts
