+++ src/internal/OpenApiTools/Walker/structure/TypeAliasNode.ts
