+++ src/internal/OpenApiTools/Walker/structure/index.ts
