+++ src/internal/OpenApiTools/components/Parameters.ts
@@ -4,8 +4,8 @@ import { Factory } from "../../TsGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Guard from "../Guard";
 import * as Name from "../Name";
-import type { Store } from "../store";
 import * as ToTypeNode from "../toTypeNode";
+import type * as Walker from "../Walker";
 import * as Paramter from "./Parameter";
 import * as Reference from "./Reference";
 import * as Schema from "./Schema";
@@ -13,7 +13,7 @@ import * as Schema from "./Schema";
 export const generateNamespace = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   parameters: Record<string, OpenApi.Parameter | OpenApi.Reference>,
   context: ToTypeNode.Context,
@@ -71,7 +71,7 @@ export const generateNamespace = (
 export const generateNamespaceWithList = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   parameters: (OpenApi.Parameter | OpenApi.Reference)[],
   context: ToTypeNode.Context,
