+++ src/internal/OpenApiTools/components/RequestBodies.ts
@@ -5,15 +5,15 @@ import { Factory } from "../../TsGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Guard from "../Guard";
 import * as Name from "../Name";
-import type { Store } from "../store";
 import * as ToTypeNode from "../toTypeNode";
+import type * as Walker from "../Walker";
 import * as Reference from "./Reference";
 import * as RequestBody from "./RequestBody";
 
 export const generateNamespace = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   requestBodies: Record<string, OpenApi.RequestBody | OpenApi.Reference>,
   context: ToTypeNode.Context,
