+++ src/internal/OpenApiTools/components/Response.ts
@@ -4,15 +4,15 @@ import type { OpenApi } from "../../../types";
 import { Factory } from "../../TsGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Name from "../Name";
-import type { Store } from "../store";
 import * as ToTypeNode from "../toTypeNode";
+import type * as Walker from "../Walker";
 import * as Header from "./Header";
 import * as MediaType from "./MediaType";
 
 export const generateNamespace = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   parentPath: string,
   name: string,
@@ -63,7 +63,7 @@ export const generateNamespace = (
 export const generateReferenceNamespace = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   parentPath: string,
   nameWithStatusCode: string,
