+++ src/internal/OpenApiTools/components/SecuritySchemas.ts
@@ -3,14 +3,14 @@ import { UndefinedComponent } from "../../Exception";
 import { Factory } from "../../TsGenerator";
 import * as Guard from "../Guard";
 import * as Name from "../Name";
-import type { Store } from "../store";
+import type * as Walker from "../Walker";
 import * as Reference from "./Reference";
 import * as SecuritySchema from "./SecuritySchema";
 
 export const generateNamespace = (
   entryPoint: string,
   currentPoint: string,
-  store: Store,
+  store: Walker.Store,
   factory: Factory.Type,
   requestBodies: Record<string, OpenApi.SecuritySchema | OpenApi.Reference>,
 ): void => {
